Permissions update

+++ app/webapi/service/user_service.go
@@ -5,6 +5,10 @@ import (
 	"district/model"
 	dto "district/model/dto"
 	repository "district/repository/handler"
+	"encoding/base64"
+	"fmt"
+	"strconv"
+	"strings"
 )
 
 type UserService struct {
@@ -15,7 +19,22 @@ func NewUserService(repositoryPool *repository.RepositoryPool) *UserService {
 	return &UserService{repositoryPool: repositoryPool}
 }
 
-func (us *UserService) GetAllUsers() ([]*dto.UserDTO, error) {
+func (us *UserService) GetAllUsers(token string) ([]*dto.UserDTO, error) {
+	decodedToken, err := base64.StdEncoding.DecodeString(token)
+	if err != nil {
+		return nil, err
+	}
+
+	tokenValues := strings.Split(string(decodedToken), ":")
+	is_admin, err := strconv.ParseBool(tokenValues[3])
+	if err != nil {
+		return nil, fmt.Errorf("your session token is not valid.")
+	}
+
+	if !is_admin {
+		return nil, fmt.Errorf("you have to be an administrator to get all users information.")
+	}
+
 	users, err := us.repositoryPool.UserRepository.GetAllUsers()
 	if err != nil {
 		return nil, err
@@ -57,19 +76,57 @@ func (us *UserService) CreateUser(request *controller.CreateUserRequest) (*dto.U
 	return dto.ConvertToUserDTO(&user), nil
 }
 
-func (us *UserService) GetUserByIdentification(identification int) (*dto.UserDTO, error) {
+func (us *UserService) GetUserByIdentification(token string, identification int) (*dto.UserDTO, error) {
 	user, err := us.repositoryPool.UserRepository.GetUserByIdentification(identification)
 	if err != nil {
 		return nil, err
 	}
 
+	decodedToken, err := base64.StdEncoding.DecodeString(token)
+	if err != nil {
+		return nil, err
+	}
+
+	tokenValues := strings.Split(string(decodedToken), ":")
+	identification, err = strconv.Atoi(tokenValues[0])
+	if err != nil {
+		return nil, fmt.Errorf("your session token is not valid.")
+	}
+	is_admin, err := strconv.ParseBool(tokenValues[3])
+	if err != nil {
+		return nil, fmt.Errorf("your session token is not valid.")
+	}
+
+	if identification != user.Identification && !is_admin {
+		return nil, fmt.Errorf("you have no permissions to check this user information.")
+	}
+
 	return dto.ConvertToUserDTO(user), nil
 }
 
-func (us *UserService) UpdateUser(identification int, request *controller.UpdateUserRequest) error {
+func (us *UserService) UpdateUser(token string, identification int, request *controller.UpdateUserRequest) error {
 	user, err := us.repositoryPool.UserRepository.GetUserByIdentification(identification)
 	if err != nil {
 		return err
+	}
+
+	decodedToken, err := base64.StdEncoding.DecodeString(token)
+	if err != nil {
+		return err
+	}
+
+	tokenValues := strings.Split(string(decodedToken), ":")
+	identification, err = strconv.Atoi(tokenValues[0])
+	if err != nil {
+		return fmt.Errorf("your session token is not valid.")
+	}
+	is_admin, err := strconv.ParseBool(tokenValues[3])
+	if err != nil {
+		return fmt.Errorf("your session token is not valid.")
+	}
+
+	if identification != user.Identification && !is_admin {
+		return fmt.Errorf("you have no permissions to update this user information.")
 	}
 
 	if request.Email != nil {
@@ -98,7 +155,31 @@ func (us *UserService) UpdateUser(identification int, request *controller.Update
 	return us.repositoryPool.UserRepository.UpdateUser(user)
 }
 
-func (us *UserService) DeleteUserByIdentification(identification int) error {
+func (us *UserService) DeleteUserByIdentification(token string, identification int) error {
+	user, err := us.repositoryPool.UserRepository.GetUserByIdentification(identification)
+	if err != nil {
+		return err
+	}
+
+	decodedToken, err := base64.StdEncoding.DecodeString(token)
+	if err != nil {
+		return err
+	}
+
+	tokenValues := strings.Split(string(decodedToken), ":")
+	identification, err = strconv.Atoi(tokenValues[0])
+	if err != nil {
+		return fmt.Errorf("your session token is not valid.")
+	}
+	is_admin, err := strconv.ParseBool(tokenValues[3])
+	if err != nil {
+		return fmt.Errorf("your session token is not valid.")
+	}
+
+	if identification != user.Identification && !is_admin {
+		return fmt.Errorf("you have no permissions to delete this user.")
+	}
+
 	return us.repositoryPool.UserRepository.DeleteUser(identification)
 }
